requirements.txt
asgiref
awesomeversion
beautifulsoup4
dj-database-url
Django==3.2.16
django-cors-headers
django-extensions
django-js-asset
gunicorn
psycopg2-binary
pytz
PyYAML
requests
six
whitenoise
django-rest-auth[with_social]
djangorestframework
markdown
django-filter
django-datatables-view
Pillow
pyimgur
SpeechRecognition
pydub
pyttsx3
cloudinary
opencv-python
qrcode